refactor: extract forms and apply modern styles
