+++ vaultpro-ui/src/components/RegisterForm.tsx
@@ -1,0 +1,83 @@
+import { useState } from 'react';
+import {
+  TextField,
+  Button,
+  Typography,
+  Select,
+  MenuItem,
+  InputLabel,
+  FormControl,
+  Paper,
+  Box,
+} from '@mui/material';
+import { useNavigate } from 'react-router-dom';
+import { register } from '../services/authService';
+
+const RegisterForm = () => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [rol, setRol] = useState('User');
+  const navigate = useNavigate();
+
+  const handleRegister = async (e: React.FormEvent) => {
+    e.preventDefault();
+    try {
+      await register(email, password, rol);
+      alert('Usuario registrado correctamente');
+      navigate('/');
+    } catch (error: any) {
+      const msg = error.response?.data?.message || 'Error al registrar usuario';
+      alert(msg);
+    }
+  };
+
+  return (
+    <Paper elevation={3} sx={{ p: 4 }}>
+      <Typography variant="h5" gutterBottom align="center">
+        Crear Cuenta
+      </Typography>
+      <Box
+        component="form"
+        onSubmit={handleRegister}
+        display="flex"
+        flexDirection="column"
+        gap={2}
+      >
+        <TextField
+          label="Email"
+          type="email"
+          fullWidth
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          required
+        />
+        <TextField
+          label="Contraseña"
+          type="password"
+          fullWidth
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          required
+        />
+        <FormControl fullWidth>
+          <InputLabel id="rol-label">Rol</InputLabel>
+          <Select
+            labelId="rol-label"
+            value={rol}
+            label="Rol"
+            onChange={(e) => setRol(e.target.value)}
+          >
+            <MenuItem value="User">Usuario</MenuItem>
+            <MenuItem value="Admin">Administrador</MenuItem>
+            <MenuItem value="Auditor">Auditor</MenuItem>
+          </Select>
+        </FormControl>
+        <Button type="submit" variant="contained">
+          Registrarme
+        </Button>
+      </Box>
+    </Paper>
+  );
+};
+
+export default RegisterForm;
