+++ vaultpro-ui/src/components/VaultList.tsx
@@ -4,7 +4,8 @@ import {
   ListItem,
   Typography,
   IconButton,
-  Box
+  Box,
+  Paper,
 } from '@mui/material';
 import DeleteIcon from '@mui/icons-material/Delete';
 import type { Contraseña } from '../services/passwordService';
@@ -15,25 +16,27 @@ interface Props {
 }
 
 const VaultList: React.FC<Props> = ({ contraseñas, onDelete }) => (
-  <List>
-    {contraseñas.map((c) => (
-      <ListItem
-        key={c.id}
-        secondaryAction={
-          <IconButton edge="end" onClick={() => onDelete(c.id)}>
-            <DeleteIcon />
-          </IconButton>
-        }
-      >
-        <Box>
-          <Typography fontWeight="bold">{c.nombre}</Typography>
-          <Typography variant="body2">
-            Usuario: {c.nombreUsuario} | Contraseña: {c.contraseña}
-          </Typography>
-        </Box>
-      </ListItem>
-    ))}
-  </List>
+  <Paper sx={{ p: 2 }}>
+    <List>
+      {contraseñas.map((c) => (
+        <ListItem
+          key={c.id}
+          secondaryAction={
+            <IconButton edge="end" onClick={() => onDelete(c.id)}>
+              <DeleteIcon />
+            </IconButton>
+          }
+        >
+          <Box>
+            <Typography fontWeight="bold">{c.nombre}</Typography>
+            <Typography variant="body2">
+              Usuario: {c.nombreUsuario} | Contraseña: {c.contraseña}
+            </Typography>
+          </Box>
+        </ListItem>
+      ))}
+    </List>
+  </Paper>
 );
 
 export default VaultList;
